+++ frontend/pages/api/v1/cmc/supply.ts
@@ -2,21 +2,21 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { PythBalance } from '@pythnetwork/staking/app/pythBalance'
 import BN from 'bn.js'
 import { STAKING_ADDRESS } from '@pythnetwork/staking/app/constants'
-import { Connection, Keypair, PublicKey } from '@solana/web3.js'
-import { Program, AnchorProvider, IdlAccounts } from '@coral-xyz/anchor'
+import { Connection, Keypair } from '@solana/web3.js'
+import { Program, AnchorProvider } from '@coral-xyz/anchor'
 import NodeWallet from '@coral-xyz/anchor/dist/cjs/nodewallet'
 import { Staking } from '@pythnetwork/staking/lib/target/types/staking'
 import idl from '@pythnetwork/staking/target/idl/staking.json'
 import { splTokenProgram } from '@coral-xyz/spl-token'
 import { TOKEN_PROGRAM_ID } from '@solana/spl-token'
-import { getConfig } from './../locked_accounts'
 import {
+  getCurrentlyLockedAmount,
+  getTotalSupply,
   getAllCustodyAccounts,
   getAllMetadataAccounts,
+  getConfig,
   getAllStakeAccounts,
-} from '../all_locked_accounts'
-
-const PYTH_TOKEN = new PublicKey('HZ1JovNiVvGrGNiiYvEozEVgZ58xaU3RKwX8eACQBCt3')
+} from '@pythnetwork/staking/app/api_utils'
 
 const connection = new Connection(process.env.BACKEND_ENDPOINT!)
 const provider = new AnchorProvider(
@@ -87,62 +87,4 @@ export default async function handlerSupply(
         "The 'q' query parameter must be one of 'totalSupply' or 'circulatingSupply'.",
     })
   }
-}
-
-function getCurrentlyLockedAmount(
-  metadataAccountData: IdlAccounts<Staking>['stakeAccountMetadataV2'],
-  configAccountData: IdlAccounts<Staking>['globalConfig']
-): PythBalance {
-  const lock = metadataAccountData.lock
-  const listTime = configAccountData.pythTokenListTime
-  if (lock.fullyVested) {
-    return PythBalance.zero()
-  } else if (lock.periodicVestingAfterListing) {
-    if (!listTime) {
-      return new PythBalance(lock.periodicVestingAfterListing.initialBalance)
-    } else {
-      return getCurrentlyLockedAmountPeriodic(
-        listTime,
-        lock.periodicVestingAfterListing.periodDuration,
-        lock.periodicVestingAfterListing.numPeriods,
-        lock.periodicVestingAfterListing.initialBalance
-      )
-    }
-  } else if (lock.periodicVesting) {
-    return getCurrentlyLockedAmountPeriodic(
-      lock.periodicVesting.startDate,
-      lock.periodicVesting.periodDuration,
-      lock.periodicVesting.numPeriods,
-      lock.periodicVesting.initialBalance
-    )
-  } else {
-    throw new Error('Should be unreachable')
-  }
-}
-
-function getCurrentlyLockedAmountPeriodic(
-  startDate: BN,
-  periodDuration: BN,
-  numPeriods: BN,
-  initialBalance: BN
-): PythBalance {
-  const currentTimestamp = new BN(Math.floor(Date.now() / 1000))
-  if (currentTimestamp.lte(startDate)) {
-    return new PythBalance(initialBalance)
-  } else {
-    const periodsElapsed = currentTimestamp.sub(startDate).div(periodDuration)
-    if (periodsElapsed.gte(numPeriods)) {
-      return PythBalance.zero()
-    } else {
-      const remainingPeriods = numPeriods.sub(periodsElapsed)
-      return new PythBalance(
-        remainingPeriods.mul(initialBalance).div(numPeriods)
-      )
-    }
-  }
-}
-
-async function getTotalSupply(tokenProgram: any): Promise<PythBalance> {
-  const pythTokenMintData = await tokenProgram.account.mint.fetch(PYTH_TOKEN)
-  return new PythBalance(pythTokenMintData.supply)
 }
