+++ frontend/pages/api/v1/locked_accounts.ts
@@ -1,18 +1,15 @@
 import { NextApiRequest, NextApiResponse } from 'next'
-import { PythBalance } from '@pythnetwork/staking/app/pythBalance'
-import BN from 'bn.js'
 import { STAKING_ADDRESS } from '@pythnetwork/staking/app/constants'
 import { Connection, Keypair, PublicKey } from '@solana/web3.js'
-import { bs58 } from '@coral-xyz/anchor/dist/cjs/utils/bytes'
-import { Program, AnchorProvider, IdlAccounts } from '@coral-xyz/anchor'
+import { Program, AnchorProvider } from '@coral-xyz/anchor'
 import NodeWallet from '@coral-xyz/anchor/dist/cjs/nodewallet'
 import { Staking } from '@pythnetwork/staking/lib/target/types/staking'
 import idl from '@pythnetwork/staking/target/idl/staking.json'
 import { splTokenProgram } from '@coral-xyz/spl-token'
 import { TOKEN_PROGRAM_ID } from '@solana/spl-token'
 import {
-  getCustodyAccountAddress,
-  getMetadataAccountAddress,
+  getStakeAccountDetails,
+  getStakeAccounts,
 } from '@pythnetwork/staking/app/api_utils'
 
 const connection = new Connection(process.env.BACKEND_ENDPOINT!)
@@ -49,114 +46,9 @@ export default async function handlerLockedAccounts(
 
     const stakeAccountDetails = await Promise.all(
       stakeAccounts.map((account) => {
-        return getStakeAccountDetails(account)
+        return getStakeAccountDetails(stakingProgram, tokenProgram, account)
       })
     )
     res.status(200).json(stakeAccountDetails)
   }
-}
-
-export async function getConfig(
-  stakingProgram: Program<Staking>
-): Promise<IdlAccounts<Staking>['globalConfig']> {
-  const configAccountAddress = PublicKey.findProgramAddressSync(
-    [Buffer.from('config')],
-    STAKING_ADDRESS
-  )[0]
-  return await stakingProgram.account.globalConfig.fetch(configAccountAddress)
-}
-
-async function getStakeAccountDetails(positionAccountAddress: PublicKey) {
-  const configAccountData = await getConfig(stakingProgram)
-
-  const metadataAccountAddress = getMetadataAccountAddress(
-    positionAccountAddress
-  )
-  const metadataAccountData =
-    await stakingProgram.account.stakeAccountMetadataV2.fetch(
-      metadataAccountAddress
-    )
-
-  const lock = metadataAccountData.lock
-
-  const custodyAccountAddress = getCustodyAccountAddress(positionAccountAddress)
-  const custodyAccountData = await tokenProgram.account.account.fetch(
-    custodyAccountAddress
-  )
-
-  return {
-    custodyAccount: custodyAccountAddress.toBase58(),
-    actualAmount: new PythBalance(custodyAccountData.amount).toString(),
-    lock: getLockSummary(lock, configAccountData.pythTokenListTime),
-  }
-}
-
-async function getStakeAccounts(connection: Connection, owner: PublicKey) {
-  const response = await connection.getProgramAccounts(STAKING_ADDRESS, {
-    encoding: 'base64',
-    filters: [
-      {
-        memcmp: {
-          offset: 0,
-          bytes: bs58.encode(Buffer.from('55c3f14f7cc04f0b', 'hex')), // Positions account discriminator
-        },
-      },
-      {
-        memcmp: {
-          offset: 8,
-          bytes: owner.toBase58(),
-        },
-      },
-    ],
-  })
-  return response.map((account) => {
-    return account.pubkey
-  })
-}
-
-export function getLockSummary(lock: any, listTime: BN | null) {
-  if (lock.fullyVested) {
-    return { type: 'fullyUnlocked' }
-  } else if (lock.periodicVestingAfterListing) {
-    return {
-      type: 'periodicUnlockingAfterListing',
-      schedule: getUnlockEvents(
-        listTime,
-        lock.periodicVestingAfterListing.periodDuration,
-        lock.periodicVestingAfterListing.numPeriods,
-        lock.periodicVestingAfterListing.initialBalance
-      ),
-    }
-  } else if (lock.periodicVesting) {
-    return {
-      type: 'periodicUnlocking',
-      schedule: getUnlockEvents(
-        lock.periodicVesting.startDate,
-        lock.periodicVesting.periodDuration,
-        lock.periodicVesting.numPeriods,
-        lock.periodicVesting.initialBalance
-      ),
-    }
-  }
-}
-
-export function getUnlockEvents(
-  startData: BN | null,
-  periodDuration: BN,
-  numberOfPeriods: BN,
-  initialBalance: BN
-) {
-  if (startData) {
-    return Array(numberOfPeriods.toNumber())
-      .fill(0)
-      .map((_, i) => {
-        return {
-          date: startData.add(periodDuration.muln(i + 1)).toString(),
-          amount: new PythBalance(
-            initialBalance.divn(numberOfPeriods.toNumber())
-          ).toString(),
-        }
-      })
-  }
-  return []
 }
